Local connection is now fixed and improved. I am fixing global connection next
git-svn-id: svn+ssh://parasol-svn.cs.tamu.edu/research/parasol-svn/svnrepository/pmpl/branches/blindRRT@3960 5e2c2758-810a-0410-8eb3-8e4e92d4004e

+++ src/Cfg/Cfg.cpp
@@ -104,7 +104,8 @@ Cfg::operator==(const Cfg& _cfg) const {
         return false;
     }
     else {
-      if(fabs(DirectedAngularDistance(m_v[i], _cfg[i])) > numeric_limits<double>::epsilon()) 
+      // decreasing numeric limits so if two Cfgs are close enough they will be equal
+      if(fabs(DirectedAngularDistance(m_v[i], _cfg[i])) > numeric_limits<float>::epsilon()) 
         return false;
     }
   }
